pyproject.toml
[project]
name = "modus-intel"
version = "0.1.0"
description = "Autonomous Threat Intelligence & IOC Enrichment Engine"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "AtlasDFIR" }]
dependencies = [
  "typer>=0.12,<1.0",
  "pydantic>=2.7,<3.0",
  "httpx>=0.27,<1.0",
  "python-dotenv>=1.0,<2.0",
]

[project.scripts]
modus-intel = "modus_intel.cli:app"

[project.optional-dependencies]
dev = [
  "black>=24.0",
  "ruff>=0.5.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]